spilt course page into multiple components
