+++ src/Component/MyButton.js
@@ -1,9 +1,10 @@
 import React from "react";
 import styled from "styled-components";
+import PropTypes from "prop-types";
 import { breakPoint } from "../utils/breakPoint";
 
 const Button = styled.button`
-    width: 100%;
+    width: ${props => props.width || "100%"};
     height: 40px;
     text-align: center;
     line-height: 40px;
@@ -26,7 +27,7 @@ const Button = styled.button`
     }
 `;
 
-export const MyButton = ({
+const MyButton = ({
     buttonWord,
     buttonId,
     clickFunction,
@@ -44,3 +45,13 @@ export const MyButton = ({
         </Button>
     );
 };
+
+MyButton.propTypes = {
+    buttonWord: PropTypes.string.isRequired,
+    buttonId: PropTypes.string,
+    clickFunction: PropTypes.func.isRequired,
+    width: PropTypes.string,
+    isDisabled: PropTypes.bool,
+};
+
+export { MyButton };
